common/types/form.tsを整理 (#94)
* common/types/form.tsのstack関連の型を移動

* common/types/form.tsのteam関連の型を移動

* common/types/form.tsのuser関連の型を移動

* common/types/form.tsのoutput関連の型を移動

* common/types/form.tsのintrospection関連の型を移動

* common/types/配下にui-elementsとui-partsフォルダ作成
