+++ context/FormContext.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, useState, useMemo } from 'react';
 import { ChildrenProps } from '@/common/types/utils';
-import { FormType } from '@/common/types/form';
+import { FormType } from '@/common/types/ui-parts/form';
 import { FormContextProps } from '@/common/types/context';
 
 const InitialState: FormContextProps = {
