+++ features/introspections/contexts/IntrospectionFormContext.tsx
@@ -1,8 +1,7 @@
 import React, { createContext, useState, useMemo } from 'react';
 import { ChildrenProps } from '@/common/types/utils';
-import { IntrospectionFormDataParams } from '@/common/types/form';
-import { IntrospectionFormContextProps } from '../types/context';
 import { InitialIntrospectionFormData } from '../functions/form';
+import { IntrospectionFormContextProps, IntrospectionFormDataParams } from '../types/context';
 
 const InitialState: IntrospectionFormContextProps = {
   introspectionFormData: InitialIntrospectionFormData,
