+++ features/introspections/types/api.ts
@@ -1,6 +1,6 @@
 import { ApiOptions } from "@/common/types/api";
+import { IntrospectionFormDataParams } from "./context";
 import { SetErrorMessages } from "@/common/types/validator";
-import { IntrospectionFormDataParams } from "@/common/types/form";
 
 export type IntrospectionApiProps = SetErrorMessages & {
   options: ApiOptions;
