+++ features/introspections/types/context.ts
@@ -1,6 +1,14 @@
-import { IntrospectionFormDataParams } from "@/common/types/form";
 import { Dispatch, SetStateAction } from "react";
-import { IntrospectionProps } from "./introspection";
+import { IntrospectionPoint, IntrospectionProps } from "./introspection";
+
+export type IntrospectionFormDataParams = {
+  reason: string;
+  evaluation: number;
+  keeps: IntrospectionPoint[];
+  problems: IntrospectionPoint[];
+  tries: IntrospectionPoint[];
+  stack_id?: number;
+};
 
 export type IntrospectionFormContextProps = {
   introspectionFormData: IntrospectionFormDataParams;
