+++ features/introspections/types/introspection.ts
@@ -1,4 +1,4 @@
-export type KeepAndProblemAndTryPoint = {
+export type IntrospectionPoint = {
   id: string;
   content: string;
   created_at: string;
@@ -9,9 +9,9 @@ export type IntrospectionProps =
   | {
       reason: string;
       evaluation: number;
-      keeps: KeepAndProblemAndTryPoint[];
-      problems: KeepAndProblemAndTryPoint[];
-      tries: KeepAndProblemAndTryPoint[];
+      keeps: IntrospectionPoint[];
+      problems: IntrospectionPoint[];
+      tries: IntrospectionPoint[];
       id?: number;
       stack_id?: number;
     }
