+++ features/outputs/comments/types/api.ts
@@ -1,7 +1,7 @@
 import { ApiOptions } from "@/common/types/api";
+import { OutputCommentFormDataParams } from "./context";
 import { SetErrorMessages } from "@/common/types/validator";
 import { SessionData } from "@/features/sessions/types/session";
-import { OutputCommentFormDataParams } from "@/common/types/form";
 
 export type CreateOutputCommentProps = SetErrorMessages & {
   options: ApiOptions;
