+++ features/outputs/comments/types/context.ts
@@ -1,5 +1,9 @@
-import { OutputCommentFormDataParams } from "@/common/types/form";
 import { Dispatch, SetStateAction } from "react";
+
+export type OutputCommentFormDataParams = {
+  content: string;
+  outputId: number;
+};
 
 export type OutputCommentFormContextProps = {
   outputCommentFormData: OutputCommentFormDataParams;
