+++ features/outputs/types/context.ts
@@ -1,5 +1,8 @@
-import { OutputFormDataParams } from "@/common/types/form";
 import { Dispatch, SetStateAction } from "react";
+
+export type OutputFormDataParams = {
+  content: string;
+};
 
 export type OutputFormContextProps = {
   outputFormData: OutputFormDataParams;
