+++ features/teams/types/context.ts
@@ -1,5 +1,14 @@
-import { InviteTeamFormDataParams, TeamFormDataParams } from "@/common/types/form";
 import { Dispatch, SetStateAction } from "react";
+
+export type TeamFormDataParams = {
+  id?: number;
+  name: string;
+};
+
+export type InviteTeamFormDataParams = {
+  id: number;
+  email: string;
+};
 
 export type TeamFormContextProps = {
   teamFormData: TeamFormDataParams;
