+++ features/users/contexts/UserFormContext.tsx
@@ -1,8 +1,7 @@
 import React, { createContext, useState, useMemo } from 'react';
 import { ChildrenProps } from '@/common/types/utils';
-import { UserFormDataParams } from '@/common/types/form';
 import { InitialUserFormData } from '../functions/form';
-import { UserFormContextProps } from '../types/context';
+import { UserFormContextProps, UserFormDataParams } from '../types/context';
 
 const InitialState: UserFormContextProps = {
   userFormData: InitialUserFormData,
